requirements.txt
2captcha-python==1.2.2
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
outcome==1.3.0.post0
packaging==23.2
pillow==10.2.0
PySocks==1.7.1
python-telegram-bot==20.8
requests==2.31.0
scipy==1.12.0
selenium==4.18.1
setuptools==68.2.2
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.27
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
wheel==0.41.2
wsproto==1.2.0
